requirements.txt
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=9.0.0
mss>=6.0.0
pyautogui>=0.9.50
openpyxl>=3.0.0
